requirements.txt
decorator==4.0.4
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
python-digitalocean==1.7
python-ping==0.0.0
requests==2.8.1
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
